use larger ints so we can handle graphs with more than 2^32 edges

+++ prototypes/thrill_louvain/src/thrill_graph.hpp
@@ -6,7 +6,7 @@
 #include <iostream>
 
 using NodeId = uint32_t;
-using Weight = uint32_t;
+using Weight = uint64_t;
 using ClusterId = uint32_t;
 
 struct Edge {
